+++ CHANGELOG.md
@@ -18,6 +18,11 @@
 - A long type-level list is broken into multiple lines. ([#665]).
 - Spaces around typed expression brackets are removed ([#666]).
 
+### Fixed
+
+- Fixed the wrong formatting of data family instances inside class instances
+  ([#667]).
+
 ### Removed
 
 - Test functions except `testAst`.
@@ -307,6 +312,7 @@ This version is accidentally pushlished, and is the same as 5.3.3.
 [@uhbif19]: https://github.com/uhbif19
 [@toku-sa-n]: https://github.com/toku-sa-n
 
+[#667]: https://github.com/mihaimaruseac/hindent/pull/667
 [#666]: https://github.com/mihaimaruseac/hindent/pull/666
 [#665]: https://github.com/mihaimaruseac/hindent/pull/665
 [#664]: https://github.com/mihaimaruseac/hindent/pull/664